requirements.txt
aniso8601==3.0.0
astroid==1.6.3
attrs==17.4.0
certifi==2018.4.16
chardet==3.0.4
colorclass==2.2.0
coverage==4.5.1
Django==2.0.4
django-cors-headers==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docopt==0.6.2
Faker==0.8.8
graphene==2.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
idna==2.6
iso8601==0.1.12
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mixer==6.0.1
more-itertools==4.1.0
packaging==17.1
pluggy==0.6.0
promise==2.1
py==1.5.3
PyJWT==1.6.1
pylint==1.8.4
pylint-django==0.11
pylint-plugin-utils==0.2.6
pyparsing==2.2.0
pytest==3.5.0
pytest-cov==2.5.1
pytest-django==3.2.1
python-bitcoinrpc==1.0
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
rope==0.10.7
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
terminaltables==3.1.0
text-unidecode==1.2
typing==3.6.4
urllib3==1.22
wrapt==1.10.11
yapf==0.21.0
